Switch to float to increase performance

+++ DynamicStiffString/DynamicStiffString.h
@@ -25,28 +25,28 @@ class DynamicStiffString
 public:
 
     struct SimulationParameters {
-        double L;
-        double rho;
-        double r;
-        double T;
-        double E;
-        double sigma0;
-        double sigma1;
+        float L;
+        float rho;
+        float r;
+        float T;
+        float E;
+        float sigma0;
+        float sigma1;
     };
 
-    DynamicStiffString(const SimulationParameters& parameters, double k);
+    DynamicStiffString(const SimulationParameters& parameters, float k);
     ~DynamicStiffString();
 
     void calculateScheme();
     void updateStates();
 
-    void refreshParameter(int changedParameterIdx, double changedParameterValue);
+    void refreshParameter(int changedParameterIdx, float changedParameterValue);
 
     void refreshCoefficients(bool init = false);
 
     //return u at the current sample at a location given by the length ratio
 
-    double getOutput()
+    float getOutput()
     {
         return v[1][6]; // set to be fixed due to varying N
     }
@@ -61,10 +61,10 @@ public:
 private:
 
     // Model parameters
-    double L, rho, r, A, T, E, I, cSq, kappaSq, sigma0, sigma1, lambdaSq, muSq, h, k;
-    double origR, origL, origE, origT, origRho;
-    std::vector<double*> parameterPtrs; // to easily locate parameters
-    std::vector<double> parametersToGoTo;
+    float L, rho, r, A, T, E, I, cSq, kappaSq, sigma0, sigma1, lambdaSq, muSq, h, k;
+    float origR, origL, origE, origT, origRho;
+    std::vector<float*> parameterPtrs; // to easily locate parameters
+    std::vector<float> parametersToGoTo;
     std::vector<bool> parameterChanged;
 
     // Number of intervals (N+1 is number of points including boundaries)
@@ -75,16 +75,16 @@ private:
     const int Mw = 1; // Mw is static
 
     // Fractional number of intervals used for dynaic grid
-    double Nfrac, NfracPrev;
-    double alf, Iterm, A0, A1, A2, A3, AA;
+    float Nfrac, NfracPrev;
+    float alf, Iterm, A0, A1, A2, A3, AA;
 
     // (N+1) x 3 'matrices' containing the state of the left and right system at all time-steps
-    std::vector<std::vector<double>> vStates;
-    std::vector<std::vector<double>> wStates;
+    std::vector<std::vector<float>> vStates;
+    std::vector<std::vector<float>> wStates;
 
     // vectors of pointers that point to state vectors
-    std::vector<double*> v;
-    std::vector<double*> w;
+    std::vector<float*> v;
+    std::vector<float*> w;
 
     /* Scheme variables
         - Adiv for u^{n+1} (that all terms get divided by)
@@ -92,16 +92,16 @@ private:
         - C for u^{n-1}
         - S for precalculated sigma terms
     */
-    double Adiv, B0, Bss, B1, B2, C0, C1, S0, S1;
+    float Adiv, B0, Bss, B1, B2, C0, C1, S0, S1;
 
     // flag to tell MainComponent whether to excite the scheme or not
     bool excitationFlag = false;
 
     // initialise location of excitation
-    double excitationLoc = 0.22;
+    float excitationLoc = 0.22f;
 
     bool clamped = true;
 
-    std::vector<double> customIp;
+    std::vector<float> customIp;
 
 };
